backend/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
pydantic-settings==2.12.0
email-validator==2.3.0
google-auth==2.45.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
pandas==2.3.3
openpyxl==3.1.5
psycopg[binary]==3.3.2
pytest==9.0.2
httpx==0.28.1
